Move json-rpc tests to a separate file
